Remove off-by-one error in set_sizes.
As an example, consider RRB_BITS = 2, meaning branching is 4. For a
dense subtrie of height 2 with 16 elements, size_sub_trie would be
called with the subtrie and parent_shift = 4. Its length would then be
calculated to be (3 << 4) + 4 = 52, which is obviously wrong.

As it should be (3 << 2) + 4 = 16, we'll just have to handle the
off-by-one error introduced.