Add size table if child has size table.
This is an issue related to slices where a newly generated node has only
one child, which may or may not contain a size table. The node doesn't
add in a size table regardless of its child. Consequently, we break the
"invariant" stating that any node N without a size table will not have
any children with a size table.

The fix is to simply add a size table when the child is an internal node
and has a size table.