Set last size_table val in sliced's root properly.
The last value in a size_table in an upcoming root will contain higher
values than it should have, when the original node did not contain a
size table and the last subtree is not completely populated. This patch
applies a fix to ensure that the last value in the size table is
correct.