Fix inconsistent rrb-tree after slice
Whenever the RRB-tree was sliced, the left slice position was 0, and
the new root was a leaf node, the root would in certain cases be not
completely filled. This lead to an inconsistency where the RRB-tree
assumed that its root was completely filled, but wasn't. This lead to
a potential segfault and an offset which returned wrong value whenever
looking up a particular index.